Return a uint64_t from NFileInfo::GetFileSize for large files when size_t is 32-bit.

+++ Library/Source/Nano/Files/NFileInfo.cpp
@@ -275,7 +275,7 @@ NTime NFileInfo::GetModifiedTime() const
 //=============================================================================
 //		NFileInfo::GetFileSize : Get the file size.
 //-----------------------------------------------------------------------------
-size_t NFileInfo::GetFileSize() const
+uint64_t NFileInfo::GetFileSize() const
 {
